# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'GESFARM API Documentation'
  description: "API de gestion d'une ferme agropastorale avec un accent sur la gestion avicole"
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
paths:
...
  /api/poultry/incubation:
    get:
      summary: 'Get Incubation Records'
      operationId: getIncubationRecords
      description: 'Retrieve incubation records'
      parameters:
        -
          in: query
          name: egg_type
          description: 'Filter by egg type'
          example: delectus
          required: false
          schema:
            type: string
            description: 'Filter by egg type'
            example: delectus
            nullable: false
        -
          in: query
          name: breed
          description: 'Filter by breed'
          example: et
          required: false
          schema:
            type: string
            description: 'Filter by breed'
            example: et
            nullable: false
        -
          in: query
          name: start_date
          description: 'date Filter by start date'
          example: quisquam
          required: false
          schema:
            type: string
            description: 'date Filter by start date'
            example: quisquam
            nullable: false
      responses:
        200:
          description: ''
          content:
            text/plain:
              schema:
                type: string
                example: "{\n  \"status\": \"success\",\n  \"data\": {\n    \"records\": [...],\n    \"pagination\": {...}\n  }\n}"
      tags:
        - 'Poultry Management'
    post:
      summary: 'Create Incubation Record'
      operationId: createIncubationRecord
      description: 'Create a new incubation record'
      parameters: []
      responses:
        201:
          description: ''
          content:
            text/plain:
              schema:
                type: string
                example: "{\n  \"status\": \"success\",\n  \"message\": \"Incubation record created successfully\",\n  \"data\": {...}\n}"
      tags:
        - 'Poultry Management'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                batch_number:
                  type: string
                  description: 'Unique batch number'
                  example: minima
                  nullable: false
                egg_type:
                  type: string
                  description: 'Egg type (chicken, duck, turkey)'
                  example: corporis
                  nullable: false
                breed:
                  type: string
                  description: 'Breed name'
                  example: eos
                  nullable: false
                egg_count:
                  type: integer
                  description: 'Number of eggs'
                  example: 12
                  nullable: false
                start_date:
                  type: date
                  description: 'Incubation start date'
                  example: aspernatur
                  nullable: false
                incubation_days:
                  type: integer
                  description: 'Incubation period in days'
                  example: 3
                  nullable: false
                temperature:
                  type: numeric
                  description: 'Temperature in Celsius'
                  example: dolorem
                  nullable: false
                humidity_percentage:
                  type: numeric
                  description: 'Humidity percentage'
                  example: debitis
                  nullable: false
                egg_size:
                  type: string
                  description: 'Egg size (small, medium, large)'
                  example: soluta
                  nullable: false
                notes:
                  type: string
                  description: 'Additional notes'
                  example: incidunt
                  nullable: true
              required:
                - batch_number
                - egg_type
                - breed
                - egg_count
                - start_date
                - incubation_days
                - temperature
                - humidity_percentage
                - egg_size
  '/api/poultry/incubation/{id}/results':
    put:
      summary: 'Update Incubation Results'
      operationId: updateIncubationResults
      description: 'Update incubation record with hatching results'
      parameters: []
      responses:
        200:
          description: ''
          content:
            text/plain:
              schema:
                type: string
                example: "{\n  \"status\": \"success\",\n  \"message\": \"Incubation results updated successfully\",\n  \"data\": {...}\n}"
      tags:
        - 'Poultry Management'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                hatched_count:
                  type: integer
                  description: 'Number of hatched eggs'
                  example: 4
                  nullable: false
                unhatched_count:
                  type: integer
                  description: 'Number of unhatched eggs'
                  example: 10
                  nullable: false
                notes:
                  type: string
                  description: 'Additional notes'
                  example: quae
                  nullable: true
    parameters:
      -
        in: path
        name: id
        description: 'Incubation record ID'
        example: 3
        required: true
        schema:
          type: integer
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'You can retrieve your token by visiting your dashboard and clicking <b>Generate API token</b>.'
security:
  -
    default: []